requirements.txt
# Core dependencies
python-telegram-bot>=22.5
langchain-text-splitters>=1.1.0
langchain-core<1.0.0
pymongo>=4.15.5
motor>=3.7.1
langchain-community>=0.4.1

# Additional dependencies
python-dotenv>=1.1.1
sentence-transformers>=2.2.2
chromadb>=0.4.17
ollama>=0.5.3

# Development dependencies
pytest>=8.3.4
pytest-mock>=3.14.0
google-generative-language>=0.1.0
